feat: 권한 상태에 따라 다른 처리가 진행되도록 기본 사항 구현
- ViewController 파일 이름 MainWeatherViewController로 변경
- guard 구문으로 significantLocationChange 기능이 가능한지 체크
- 위치 권한이 결정이 되지 않아도 일단 추적을 하도록 로직 구현
- 처음 locationManager가 생성되었을때에도 locationManager(_:didChangeAuthorization:)이 호출되므로 권한과 관련된 사항은 모두 해당 메서드 내에 작성
