+++ WeatherForecast/WeatherForecast/Controller/MainWeatherViewController.swift
@@ -1,0 +1,38 @@
+//
+//  MainWeatherViewController - ViewController.swift
+//  Created by yagom. 
+//  Copyright © yagom. All rights reserved.
+// 
+
+import UIKit
+import CoreLocation
+
+class MainWeatherViewController: UIViewController {
+    let locationManager = CLLocationManager()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        locationManager.delegate = self
+        guard CLLocationManager.significantLocationChangeMonitoringAvailable() else { return }
+        locationManager.startMonitoringSignificantLocationChanges()
+    }
+}
+
+extension MainWeatherViewController: CLLocationManagerDelegate {
+    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+        switch status {
+        case .restricted, .denied:
+            break
+        case .authorizedAlways, .authorizedWhenInUse:
+            break
+        case .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
+        @unknown default:
+            break
+        }
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+    }
+}
+
